Rework component provider to make it easier to compose
