+++ src/core/entities/PollResult.ts
@@ -2,8 +2,13 @@ import { Answer } from './Answer'
 import { UserConsentData } from './UserConsentData'
 import { UserProfile } from './UserProfile'
 
-export interface PollResult {
+export interface PollRemoteQuestionResult {
   answers: ReadonlyArray<Answer>
+}
+
+export interface PollUserInformationsResult {
   profile: UserProfile
   consentData: UserConsentData
 }
+
+export type PollResult = PollRemoteQuestionResult & PollUserInformationsResult
